finish quad tree extract function
